requirements.txt
# Core Django
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.2.1

# Optional shared cache / throttling
django-redis==6.0.0
redis==7.3.0

# Utilities
cryptography>=42,<45
pytz==2024.1
python-dotenv==1.0.1
holidays==0.89
icalendar==6.0.1
recurring-ical-events==3.3.2
Pillow==10.3.0
pandas>=2.0
openpyxl>=3.1
vercel==0.5.8
